refactor(ui): extract ZeroIntrinsicWidth/Height into one shared SSOT
MusicCard's ZeroIntrinsicHeight and SpeedOverlay's ZeroIntrinsicWidth
were the same MeasurePolicy scaffold with only the zeroed axis differing.
Consolidate them into ui/home/components/ZeroIntrinsicSize.kt: one
ZeroIntrinsicSizePolicy(zeroWidth) backs both public wrappers, which keep
their axis-named call sites. Delegating the non-zeroed axis to the child's
own intrinsic matches the prior MeasurePolicy default, so behaviour is
unchanged (SpeedOverlayWidthTest, MusicCardHeightTest, and all dashboard
goldens unchanged). Addresses a PR #313 review note.
